% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbcAnalysisTemplate-package.R
\docType{package}
\name{sdbcAnalysisTemplate-package}
\alias{sdbcAnalysisTemplate}
\alias{sdbcAnalysisTemplate-package}
\title{sdbcAnalysisTemplate: Rmd Template for Statistical Analysis}
\description{
Stores custom formatting and html structure that I use a lot when doing exploratory analysis in R.
}
\author{
\strong{Maintainer}: Emma Braun \email{emma.braun@hsc.utah.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
